Append file type to filename on image upload
Previously image uploads determined the final file path to upload by
appending the file type suffix to the file path prior to calling glance
upload. This broke recently when we fixed a different bug in glance
uploads because we stopped updating the filename var with the suffix.
Add this update back in.

Change-Id: I8263f5389c3b76d5b55c12913af98f7b1cc0a42f