Fix exception in dib image cleanup
Nodepool might start an image build and it might fail between
the top of the periodicCleanup method and the bottom.  Yet
the method keeps a list of dib image ids that it reuses.  If
an image build failed, that dib image may no longer exist, so
if the DB does not return a row for it, don't attempt to use it.

Change-Id: I47113e89fbb1dc45d8fa96ddabce421f80d800a9