Be more careful when config options are not set
Currently, when the config contains no definition for elements-dir
or script-dir, there will be a pretty obscure error like

TypeError: execve() arg 3 contains a non-string value

when disk-image-builder is executed, because the environment
variable is set to None. The solution is to avoid setting the variable
in this case completely.

Also try to copy scripts from scriptdir only if it was set.

Change-Id: Id683a57b15dcb0974c9a58070e971f7ea0dc9b99