Add hostname templates to config validator
In I5d5650fe6b22ecb25b994767e48e7742d7238a18 some new configuration
options were added in order to be able to change the hostname that
is being set for new nodes. Now we add them to the validation code,
too, so that the config can in fact be validated again.

Change-Id: I4f849fcd674447b26c4508018af5633ae2142b2e