Remove terminated image/instance delete threads
When the separate image and instance delete threads were added
for periodic cleanup of leaked images and instances, the same
anti-duplication mechanism used for the periodic cleanup of nodes
was used: an dictionary of thread objects keyed by the object
being deleted.  However, the cleanup of those delete threads
was neglected, meaning that after those threads exited, if the
object persisted, nodepool would never try to delete it again.

This change adds the missing cleanup code.

Change-Id: Iac4049d265e23a7eae1b2917181de3c35fb6359d