Add comment field to node and reason to hold
When holding a node, there is almost always a reason. However, it's hard
to communicate that reason to other nodepool admins. Add a comment field
to the node table and a --reason argument to the hold command that allow
for annotating a held node with the reason why the node was held. This
should allow, if nothing else, to let people know who to talk to about a
held node.

NOTE: This will require a database schema alteration to add the comment
column.

Change-Id: I31880745e97b4fbd98ad5ec0133c1682749dfb1e