Bust cache in leaked node test
The leaked node test relies on listing nodes from the provider
to determine that leaked nodes have been removed.  These listings
are cached.  It is possible for the cache to not be updated by
the time the assertion at the end of the test runs, so whenever
the test directly queries the provider for its instance listing,
ignore the cache.

Change-Id: Iddeacdc042d559a1cb32f8893b7f6a5deb398da8