Stop nodepool once in tests
The test for dib and snapshot image build fails uses the useNodepool
utility function which sets up a cleanup to stop the nodepool thread for
us so we can remove the explicit nodepool stopping in this test.

Do this as double stopping is generally something you should avoid and
has impacted this test.

Change-Id: I7135673b78401c7f04c0154967fe29f9a9b90ed6