Permit obsolete jenkins/mysql options in config
Nodepool no longer recognizes these config options, however, the
transition strategy involves having the new config on disk before
restarting an upgraded nodepool.  It is convenient to allow the
configuration to exist in both places so that the configuration
update and nodepool upgrade/restart do not need to be synchronized.

After a short time, we can remove these entries from validation.

Change-Id: Ia646321605dd72db26b7b3c5bd6fd92afbc03889