Explicitly block in nodepool upload commands
For the image upload/update commands we do not actually block
explicitly. The commands do not return until the spawned thread is
complete because the thread is a non-daemon thread which causes the
python interpreter to wait for their completion.

Because we want these commands to block, lets make them do so
explicitly.

Change-Id: I96c88ecfeb3a2836d7b6069d5e92314a30f251e1