Don't double-log in nodepool/dsvm log output
Currently we are double-logging lines in the nodepoold output during
dsvm tests with the configuration file added in
I5113898bf57d1f365ad345b772606feec7c9e818

We do not wish to propagate the log messages from the "nodepool.*"
handler; since it's logging at the lowest level, all messages have
already been handled.

<tmi>

You'd think that the "level=WARNING" in [logger_root] would filter
this out, and thus we wouldn't see duplicates of < WARNING level
messages.  But we do...

But there's a trick/trap though when reading [1], in that when the
log-message propagates upwards from [logger_nodepool] it is just
passed to the *handlers* of the parent logger -- not the parent logger
itself; i.e. the filtering in [logger_root] does not apply and we
essentially pass the message to the "console" handler twice.  Hence
the double-logging.

</tmi>

[1] https://docs.python.org/2/howto/logging.html#logging-flow

Change-Id: I627c3575a3b43ba87b5e450a9bafbace75bfce32