Get public IPs from supplied networks
If the user specifies a net-label, try to find the public IP
on the corresponding network.  NB: this only works for net-label,
not net-id, because nova returns the network dictionary keyed by
the label; to use the ID would require a separate lookup.

This allows us to find public IPs on Internap.  If we switched
HPCloud to using net-label instead of net-id, it may work there
as well (though this is untested).

Change-Id: I616f851b3e3bac817375a7cf0511845703b749f7