Fail on bad options to fake-image-create
fake-image-create currently just ignores options it doesn't know
about.  I'd rather see us being strict about this, so if we introduce
typos, etc, we find it.  Expand the getopt call for the extra args
passed in currently.

(if this gets too complicated, a quick re-write in python with
argparse would make things easier)

Change-Id: If1eb1774549463644fc1d667d17773c4e30cbb60