Partially Revert "Pass extended network information in to occ/shade"
Adding the config value "nat_destination" increased the surface area of
the nodepool.yaml networks list being used for two reasons. The primary
purpose of nodepool.yaml is to tell nodepool what to create - and the
networks list says "please boot the vms for this provider on all of the
networks in this list". However, there are, for some clouds, additional
information that needs to be configured about the cloud that should not
also imply that nodes should boot on the associated networks. A good
example is internap, where the networks do not follow patterns that
allow them to be discovered. One needs, in config, to tell shade "this
network is public and this network is private" but nodepool really only
wants to boot with a nic connected to the public network.

Long story short - we should not aim to duplicate all of the features of
clouds.yaml in nodepool.yaml. They are two different files with two
different purposes in life. One should use clouds.yaml to configure
qualities of the cloud, and nodepool.yaml to configure how you want to
use the cloud.

This reverts commit 2a30810b2e111e66f84bc65b20e2b050f485ed1c.

Change-Id: Ic7801d2ad49574f2b24c28c5a8f747ba73c6bf24