Infer node demand from workers listing
We want to support dropping the labels off of the job names, but current
node demand calculations rely on the labels being in the job names. The
good news is that nodepool knows the type/label of instances because it
booted them so we can create this mapping using the workers admin
command which returns worker to functions listings.

Change-Id: Ic45081e0f9560821194af9e7db179f77c2288d53