Include subnodes in updateStats function v2
When we plot data in graphite / grafana it is lacking info about our
subnodes. This doesn't give us a true reflection of the nodes
currently building / ready / in use / delete.

Previously we queried SQL for our total nodes calculation. We suspect
that was an expensive task which caused problems in production. As a
result, we now look to the labels configuration to get the
information.

Change-Id: Icd4f6ce8c50df563ea2faa198703f5e2892dbc62
Signed-off-by: Paul Belanger <pabelanger@redhat.com>