Have buildZooKeeperHosts accept a config object
Add a new ZooKeeperServerConnection class which represents the
information needed to connect to a server, and then have the
config parser create that object instead of its own internal
representation of the connection.  Modify the ZooKeeper class
to accept this new object to specify what connection(s) it
should use.

The new class implements the equality operator so that it may
still be used in the same way as other ConfigValue objects (that
is, checking to see if the configuration has been updated).

Change-Id: I4cfa3866e3b0ae2a730c1e624570cd7025901808