Change ZK API to not auto-create build on lock
One of the current lock context managers will automatically create
a new build node for the caller, the other does not. It turns out,
after beginning implementation of the build logic, that this auto
creation is not a good idea and causes problems. There are usually
checks one needs to do for validation before deciding to actually
begin the build. Also, the existence of the build node with no
data associated with it while the build is happening causes other
code to have to work around it.

Instead, just have a single lock context manager that does NOT
create a new node and change storeBuild() to create a new node for
the caller if the build number is not supplied.

Change-Id: Ifbf6d73efff0ba32eb5f21d72cb63d677f109b67