Short-circuit builder processing on shutdown
The builder will properly abort a current image build if it has
been told to shutdown. However, it will continue cycling through
the remaining images until control returns to the main run loop.
We don't want that, so add short-circuits that checks to see if
we should still be running.

Change-Id: I44b1ca8fd3adacc1b5581f933d61863398ea0f92