Add provider upload framework
Adds the logic around uploading local DIB image builds to the
providers that need them.

This code expects that the local images will be named using the
sequence ID of the build, and that the build data, as stored in
the ZooKeeper node, will contain the formats available for the
built image. Storing the formats lets us handle the case when a
new provider may have been added to the config file that needs a
new image format. If the format is not available for upload to the
new provider, it will just wait for the next build which should
have the new format.

Change-Id: Id6f98f5d64be6ab72dd0c38a452ab7ad09a6ba3b