Add getMostRecentBuildImageUpload method to zk
Much of nodepool is going to want to use a method that returns
the most recent image upload of an image in general, rather
than the most recent upload of a specific build of an image.

The natural name for that method would be "getMostRecentImageUpload"
so this change renames the current method with that name to
"getMostRecentBuildImageUpload" to indicate that it expects a build
parameter.  "getMostRecentImageUpload" is then created to return
the most recent upload given only an image and provider name.

Change-Id: Ib6c14ece4daf2e72059bba927a450854cdefa336