Assume diskimage and image names are the same
The nodepool ZK spec assumed that there would be a 1:1 mapping
between diskimages and provider images.  That is to say that
a given diskimage "foo" would not be uploaded to a provider twice
to supply "foo-a" and "foo-b" images for nodepool.  In practice,
we observe this 1:1 relationship.  However, nodepool (and in
particular, the config file format) allows for a diskimage to be
re-used like this.

We need to correct nodepool and the config file format to enforce
the 1:1 mapping that we desire.  However, in the mean time, this
change will allow us to use our current test fixtures by building
the assumption that there is a 1:1 correspondence between provider
images and diskimages into the builder code.  We can later alter
the config file format to eliminate this.

With this change, when the builder needs to find the provider image
record for a given diskimage, it does so by searching for the
diskimage name in the provider image records, rather than assuming
that the image name is the same as the diskimage name.

Change-Id: Ibe2e91555400d788b7cd2991e3eec0521291b0b4