Add option to force image delete
If the upstream cloud is offline or you have forgotten to remove
images before removing a provider in the config, you can't delete the
images associated with it in nodepool, since trying to delete the
upstream image fails.

This adds a "--force" option that tells nodepool we understand that
the remote image will not be removed but to go ahead and cleanup
locally anyway.

Change-Id: I416cb21346b6c0eb9932bbfc7bf67db2b9ac7ee3