Remove snapshot support
We don't have plans to support building images from snapshots in the
future.  If we add it back, it will require a new builder implementation
(with zookeeper interface, etc) anyway.

In addition to removing essentially dead code, this also begins the
process of simplifying the configuration file.  It removes 'base-image'
and 'setup' from the provider image attributes since they are only used
for snapshot images.

Change-Id: I68e0d2c7017cfebe35341189b5aedb0556bcd78b