Activate virtualenv before running dib
When nodepool-builder is run directly from a virtualenv
(/path/to/env/bin/nodepool-builder) the $PATH does not have the
virtualenv's "bin" in it.  Adding that directory to $PATH is basically
all that "activating" a virtualenv does (e.g. sourcing activate.sh).

Since nodepool-builder just wants to call "disk-image-create" and get
the version that was installed along-side it from requirements.txt, we
either need to enforce the rule that you only run nodepool from an
activated virtualenv, or, we can help out with activation as proposed
here.

virtualenv includes some python to activate the virtualenv within the
current process.  We hook that into a pre-exec function to run before
we call disk-image-create.

This keeps dib happy too -- it needs to be run in an activated
virtualenv for basically the same reason; it looks for companion tools
like dib-run-parts (dib should actually do something like this too for
robustness but this is still necessary here so we can just call
"disk-image-create" with no prefix).

The upshot is, systemd start scripts should be able to just call
/path/to/env/bin/nodepool-builder and it will "just work" without them
having to do tricks to activate the virtualenv before daemon start.

Change-Id: I0537cbf167bb18edf26f84ac269cbd9c8a1ea6a2