Fix race in test_image_rotation
We have a very rare failure: waiting for build 001 to be deleted
sometimes fails in test_image_rotation. This is because if things
happen quickly enough, builds 001 and 002 will get the same upload
state_times, which can throw off the CleanupWorker's calculation
of "most recent uploads" (it would think 001 is more recent than 002,
due to sorting on state_time). Add a slight delay to make sure they
have different times.

Change-Id: I2dc127b7cd42d3ee7afb1f0677810554a800e8a8