Add pause option for image uploads
We need to ability to pause uploads of our images to a providers. For
example, we've build a good image, but it includes a regression for a
specific provider.

By setting the pause option of the image, we are able to delete the
bad image from the provider and prevent nodepool-builder from
re-uploading it.

Change-Id: Ib03852ea378fcdc985825b2e477036b06c53ea57
Signed-off-by: Paul Belanger <pabelanger@redhat.com>