Fix image-delete command
On manual image delete requests for an image build that is current,
we need to actually delete the upload from the provider, then remove
the ZK record. To do so, we have to call _deleteUpload() directly.
Rather than go through the manipulations to pass that method the
parameters it needed (e.g., a provider object), turns out that the
upload object has all of the information that _deleteUpload() actually
needs, so clean that up here.

Also fixes the test_image_delete test to meet expectations.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>
Change-Id: Ib0e4f0f194504b1b430078102e5b8d80b6a015fb
Story: 2000812
Task: 3397