Validate configs when used by tests
We have somewhat frequently failed to update our voluptuous schema when
adding new content to our config because the tests for it are in a
corner. Address this by testing that every test's config validates
properly when a test applies a config. This means that whenever we add
tests for a new feature it needs to have working config validation too.

Change-Id: Ie4452747baaf3d89d51da8f252366a6919f4d10a