Mitigate config disagreement issues
Because the config file can change on us at any point, and each thread
carries around its own copy of it, it's possible that a thread can be
using an outdated version while in the middle of its work. This can
cause unexpected behavior, like uploading an image to a provider when
it is currently being deleted (leaving an undeleted upload). Until we
can solve the deeper issue of config file syncing, we can reduce
probability of this delete/upload problem by adding a check for the
build status _just_ before an upload attempt, and by reloading our
config after we spent time trying to upload to a provider (which can
take quite a while).

Also removes a TODO that was TODONE.

Change-Id: Id066361792d356cf7f22b9157092e364a848c8e1