Ensure env-vars are strings in config validate
If a non-string value is passed to env-vars then when diskimage-builder
is run it will be injected as is into the ENV dictionary. This results
in the error "TypeError: execve() arg 3 contains a non-string value"
which is not exactly explicit about the problem.

We have two options here: Either convert the dictionary items to strings
automatically or make the config file fail validation. For now I'm going
with config validation because if the value is a dict or something more
difficult than simply an int this could be confusing.

Change-Id: I94013b50761ff0fc5569eb1af6420454b4191a50