Make web status text/plain
When I added the JSON targets in
I3410a4e4efd649732747b0502d66d212f50fa1bb I also changed the text
output mime-type to be application/text, which seemed to pair with
application/json.  But this means browsers prompt to download the file
rather than display it, which is not the intention.  Restore to
text/plain.

Change-Id: I6fa6597b2fcd05a4c33ba4932c966101daf98ecf