Add 'hostname-format' to provider config section
We will no longer get the hostname format from the 'target' section
of the config file, and instead just grab it from the 'provider'
section. Move 'hostname' to 'hostname-format' in the 'provider' section.

Also, rename 'template-hostname' to 'image-name-format' since it's
used for the external (provider) image names and not really hostnames.

Change-Id: I1ab2644d31eb371e1cf6b5e67ed2d2f850cf2464