Fix potential race in image upload cleanup
It's possible for an active image upload to complete during the
builder cleanup phase, which would release its lock and change
the state to READY. If the cleanup thread has cached the UPLOADING
state and checks to see if the upload is in progress (just after
the lock has been release), then we could accidentally delete the
upload record because we don't have the correct state. This adds
a second check on the state to make sure it didn't change on us.

Change-Id: I60be43e999cf86d4c3c46e6ea69ecd1bcb69f533