Don't try and delete nodes with no external_id
When a node upload fails it ends up with a node in zookeeper with an
empty external_id. The cleanup thread then tries to delete the None ID
from OpenStack. If there is no external_id and the node is up for
deletion then skip the remote delete part and just remove the zookeeper
node.

Change-Id: If750f108011c04ab16d0316094b8d1133a083e9b