Fix for unpaused request handlers
We've been seeing some random test failures where paused handlers
never unpause. I believe this may be the cause. When looping through
the request's node types, we never took into consideration nodes that
we've already put into our node set (if it had paused to wait for
nodes). This would cause the handler code to try to grab more nodes
than was required to satisfy the request. Since some of the tests
limit max-servers to a very low number, this could cause the test to
hang.

Change-Id: Ifb87563061de152ee2407b02845044ab06648a7c