Update wait_for_threads
We can remove several singleton threads from the wait_for_threads
whitelist since they no longer exist.  NodeCleanupWorker is a singleton
thread so can be removed from the pattern match section to the whitelist.

NodeLauncher-* threads are not a long-running, so we don't need to wait
for them.  In fact, that we were waiting on them was likely making the
tests racier than necessary, since the intent here is to wait for
Nodepool to stabilize as much as possible before proceeding to the next
phase of a test.

Change-Id: I760fc445404578671be7284532f222fb2cf4c991